[front/refactoring] #6, favicon
