+++ frontend/package.json
@@ -35,6 +35,7 @@
     "react-quill": "^1.3.3",
     "react-redux": "^7.0.3",
     "react-responsive": "^6.1.2",
+    "react-reveal": "^1.2.2",
     "react-router-dom": "^4.4.0-beta.8",
     "react-scripts": "3.0.1",
     "recompose": "^0.30.0",
