+++ frontend/src/pages/main-page/styles.scss
@@ -14,6 +14,7 @@
 
 .main-page-wrapper {
   padding: 100px 10vw;
+  min-height: 1000px;
 
   .blog-item-card {
     padding: 50px 30px 0 30px;
