+++ frontend/tsconfig.json
@@ -15,5 +15,6 @@
     "noEmit": true,
     "jsx": "preserve"
   },
-  "include": ["src"]
+  "include": ["src", "typings", "node_modules/@types"],
+  "exclude": ["node_modules"]
 }
