+++ frontend/typings/react-reveal.d.ts
@@ -1,0 +1,2 @@
+declare module 'react-reveal/Reveal'
+declare module 'react-reveal/Fade'
